requirements.txt
annotated-types==0.7.0
anyio==4.8.0
av==14.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.5.0
distro==1.9.0
enum34==1.1.10
exceptiongroup==1.2.2
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2025.2.0
groq==0.18.0
h11==0.14.0
halo==0.0.31
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
log-symbols==0.0.14
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.20.1
openwakeword==0.6.0
packaging==24.2
protobuf==5.29.3
pvporcupine==1.9.5
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
RealtimeSTT==0.3.94
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
spinners==0.0.24
sympy==1.13.1
termcolor==2.5.0
tflite-runtime==2.14.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==14.2
flask==3.0.0
flask-sock==0.3.2
flask-cors==4.0.0
opencv-python~=4.10.0.84
ultralytics~=8.3.73